requirements.txt
python-dotenv
pandas
numpy==1.23.5
scipy==1.9.3
scipy
ipykernel
pymongo
certifi
pymongo==4.10.1
pymongo[srv]
scikit-learn==1.2.2
mlflow
pyaml
# dill
dagshub

fastapi
uvicorn
python-multipart

xgboost


##-e .